Significantly improved SVM and fully integrated classifier into the project (~93% correctness); implemented radix sort for ordering points to improve efficiency.
-The ij finger/defect/centroid coordinates in Hand and the contour points
in Object3D are no longer 2x their real values on screen and are automatically
scaled to the correct size during computation.

-Created a new pipeline to get information from DepthCamera: operator()
for updating, >> for reading images & objects (not sure if
will keep this in an actual release?)

-NOTE: two commits ago I actually changed how the DepthCamera works
significantly, and forgot to document it.
The old way of using computeClusters()/getClusters()/for ... Object3D() still works but
now we can simply call queryObjects() or even queryHands() to do this in one go.
Also, to update the camera, call nextFrame() instead of update()
(which is now protected and called by nextFrame) to automatically remove noise, clear stored information, etc.
This makes using OpenARK from the outside a lot simpler.

-Visualization: now displays the IR image behind the detected contours if available. This is
makes it easier to identify which part of the object/wall we are detecting
-The visualization now shows the lengths each fingers in cm above the line marking it.
-The points, lines, and text drawn by the hand visualizer now automatically scale with camera image size.

-No longer using IR information in clustering due to lack of effectiveness & small but appreciable performance impact
-Made minor changes to the way TestCamera works to fit with the new nextFrame() pipeline and also to improve reading performance

-Using pointers instead of at<>() everywhere to improve performance.

-Implemented A* shortest path in Util to check how far two fingers are
if only going through points in cluster, though decided against using as a discriminant because was too slow and prone to noise.